fix: 修复 .eslintrc.json 结果被保存到 .eslintrc

+++ src/scripts/util.js
@@ -63,7 +63,11 @@ function readWoodpeckerEslintCfg() {
 }
 
 function writeEslintCfg(newEslintCfg) {
-  fs.writeFileSync(path.resolve(cwd, './.eslintrc'), JSON.stringify(newEslintCfg, null, 2))
+  if (cfgFileInfo.eslintCfgExist) {
+    fs.writeFileSync(cfgFileInfo.eslintCfgPath, JSON.stringify(newEslintCfg, null, 2))
+  } else {
+    fs.writeFileSync(path.resolve(cwd, './.eslintrc'), JSON.stringify(newEslintCfg, null, 2))
+  }
 }
 
 function readPrettierCfg() {
